+++ examples/numba/example06_trainingtimevslightgbm.py
@@ -16,93 +16,60 @@
    https://github.com/elephaint/pgbm/blob/main/LICENSE
 
 """
-#%%
-import torch
+#%% Import packages
+from pgbm_nb import PGBM
+import numpy as np
 import time
-from pgbm import PGBM
 from sklearn.model_selection import train_test_split
 import lightgbm as lgb
 from datasets import get_dataset, get_fold
-import numpy as np
 #%% Objective for pgbm
-def mseloss_objective(yhat, y, levels=None):
-    gradient = (yhat - y)
-    hessian = torch.ones_like(yhat)
-
-    return gradient, hessian
-
-def rmseloss_metric(yhat, y, levels=None):
-    loss = (yhat - y).pow(2).mean().sqrt()
-
-    return loss
-
-def mseloss_objective_np(yhat, y, levels=None):
+def mseloss_objective(yhat, y):
     gradient = (yhat - y)
     hessian = np.ones_like(yhat)
 
     return gradient, hessian
 
-def rmseloss_metric_np(yhat, y, levels=None):
-    loss = np.sqrt(np.mean(((yhat - y)**2)))
+def rmseloss_metric(yhat, y):
+    loss = np.sqrt(np.mean(np.square(yhat - y)))
 
     return loss
 #%% Load data
 #datasets = ['boston', 'concrete', 'energy', 'kin8nm', 'msd', 'naval', 'power', 'protein', 'wine', 'yacht','higgs']
-dataset = 'msd'
+dataset = 'protein'
 data = get_dataset(dataset)
 X_train, X_test, y_train, y_test = get_fold(dataset, data, random_state=0)
 X_train_val, X_val, y_train_val, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)
 #%% Parameters
-base_estimators = 100
+base_estimators = 500
 params = {'min_split_gain':0,
       'min_data_in_leaf':1,
-      'max_leaves':8,
-      'max_bin':64,
+      'max_leaves':32,
+      'max_bin':256,
       'learning_rate':0.1,
       'n_estimators':base_estimators,
       'verbose':2,
-      'early_stopping_rounds':100,
+      'early_stopping_rounds':base_estimators,
       'feature_fraction':1,
       'bagging_fraction':0.1,
       'seed':1,
       'lambda':1,
       'tree_correlation':0.03,
-      'device':'gpu',
-      'output_device':'gpu',
-      'gpu_device_ids':(0,),
-      'derivatives':'exact',
+      'split_parallel':'feature',
       'distribution':'normal'}
-#%% PGBM
-# Tuples of datasets in tensors
-train_val_data = (X_train_val, y_train_val)
-valid_data = (X_val, y_val)
-# Train 1 iteration to compile code
-params['n_estimators'] = 1
-model = PGBM()    
-model.train(train_set=train_val_data, objective=mseloss_objective, metric=rmseloss_metric, valid_set=valid_data, params=params)
-# Train to retrieve best iteration
-params['n_estimators'] = base_estimators
-start = time.perf_counter()
-model = PGBM()    
-model.train(train_set=train_val_data, objective=mseloss_objective, metric=rmseloss_metric, valid_set=valid_data, params=params)
-torch.cuda.synchronize()
-end = time.perf_counter()
-print(f'Fold time: {end - start:.2f}s')
 #%% PGBM-Numba
-# Import numba-backend package
-from pgbm_nb import PGBM
 # Tuples of datasets in tensors
 train_val_data = (X_train_val, y_train_val)
 valid_data = (X_val, y_val)
 # Train 1 iteration to compile code
 params['n_estimators'] = 1
 model = PGBM()    
-model.train(train_set=train_val_data, objective=mseloss_objective_np, metric=rmseloss_metric_np, valid_set=valid_data, params=params)
+model.train(train_set=train_val_data, objective=mseloss_objective, metric=rmseloss_metric, valid_set=valid_data, params=params)
 # Train to retrieve best iteration
 params['n_estimators'] = base_estimators
 start = time.perf_counter()
 model = PGBM()    
-model.train(train_set=train_val_data, objective=mseloss_objective_np, metric=rmseloss_metric_np, valid_set=valid_data, params=params)
+model.train(train_set=train_val_data, objective=mseloss_objective, metric=rmseloss_metric, valid_set=valid_data, params=params)
 end = time.perf_counter()
 print(f'Fold time: {end - start:.2f}s')
 #%% LightGBM