+++ examples/numba/example07_optimizeddistribution.py
@@ -1,0 +1,92 @@
+"""
+   Copyright (c) 2021 Olivier Sprangers 
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+
+   https://github.com/elephaint/pgbm/blob/main/LICENSE
+
+"""
+#%% Import packages
+import numpy as np
+from pgbm_nb import PGBM
+from sklearn.model_selection import train_test_split
+from datasets import get_dataset, get_fold
+#%% Objective
+def objective(yhat, y):
+    gradient = (yhat - y)
+    hessian = np.ones_like(yhat)
+    
+    return gradient, hessian
+
+def rmseloss_metric(yhat, y):
+    loss = np.sqrt(np.mean(np.square(yhat - y)))
+
+    return loss
+#%% Generic Parameters
+# PGBM specific
+params = {'min_split_gain':0,
+      'min_data_in_leaf':1,
+      'max_leaves':8,
+      'max_bin':64,
+      'learning_rate':0.1,
+      'n_estimators':100,
+      'verbose':2,
+      'early_stopping_rounds':100,
+      'feature_fraction':1,
+      'bagging_fraction':1,
+      'seed':1,
+      'lambda':1,
+      'tree_correlation':0.03,
+      'split_parallel':'feature',
+      'distribution':'normal'}
+n_forecasts = 1000
+#%% Loop
+# datasets = ['boston', 'concrete', 'energy', 'kin8nm', 'msd', 'naval', 'power', 'protein', 'wine', 'yacht']
+dataset = 'boston'
+# Get data
+data = get_dataset(dataset)
+X_train, X_test, y_train, y_test = get_fold(dataset, data, 0)
+X_train_val, X_val, y_train_val, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=0)
+# Build datasets
+train_data = (X_train, y_train)
+train_val_data = (X_train_val, y_train_val)
+valid_data = (X_val, y_val)
+# Set base number of estimators
+base_estimators = 2000
+params['n_estimators'] = base_estimators
+# Train to retrieve best iteration
+model = PGBM()
+model.train(train_val_data, objective=objective, metric=rmseloss_metric, valid_set=valid_data, params=params)
+# Find best tree correlation & distribution on validation set
+best_distribution, best_tree_correlation = model.optimize_distribution(X_val, y_val)
+#%% Train model on full dataset and evaluate base case + optimal choice of distribution and correlation hyperparameter
+# Set iterations to best iteration from validation set
+params['n_estimators'] = model.best_iteration + 1
+# Retrain on full set   
+model = PGBM()
+model.train(train_data, objective=objective, metric=rmseloss_metric, params=params)
+#% Probabilistic predictions base case
+model.params['distribution'] = 'normal'
+model.params['tree_correlation'] = 0.03
+yhat_dist_pgbm = model.predict_dist(X_test, n_forecasts=n_forecasts)
+# Scoring
+crps_old = np.mean(model.crps_ensemble(yhat_dist_pgbm, y_test))
+# Optimal case
+model.params['distribution'] = best_distribution
+model.params['tree_correlation'] = best_tree_correlation
+yhat_dist_pgbm = model.predict_dist(X_test, n_forecasts=n_forecasts)
+# Scoring
+crps_new = np.mean(model.crps_ensemble(yhat_dist_pgbm, y_test))  
+# Print scores
+print(f"Base case CRPS {crps_old:.2f}, distribution = normal, tree_correlation = 0.03")
+print(f"Optimal CRPS {crps_new:.2f}, distribution = {model.params['distribution']}, tree_correlation = {model.params['tree_correlation']}")