+++ examples/pytorch/example07_optimizeddistribution.py
@@ -20,8 +20,6 @@
 import torch
 from pgbm import PGBM
 from sklearn.model_selection import train_test_split
-import pandas as pd
-import numpy as np
 from datasets import get_dataset, get_fold
 #%% Objective
 def objective(yhat, y, levels=None):
@@ -60,7 +58,6 @@ n_forecasts = 1000
 #%% Loop
 # datasets = ['boston', 'concrete', 'energy', 'kin8nm', 'msd', 'naval', 'power', 'protein', 'wine', 'yacht']
 dataset = 'boston'
-df_val = pd.DataFrame(columns=['method', 'dataset','fold','device','validation_estimators','test_estimators','rho','distribution','crps_validation'])
 # Get data
 data = get_dataset(dataset)
 X_train, X_test, y_train, y_test = get_fold(dataset, data, 0)
@@ -76,17 +73,7 @@ params['n_estimators'] = base_estimators
 model = PGBM()
 model.train(train_val_data, objective=objective, metric=rmseloss_metric, valid_set=valid_data, params=params)
 # Find best tree correlation & distribution on validation set
-tree_correlations = np.arange(10) * 0.01
-distributions = ['normal','studentt','laplace','logistic','lognormal', 'gumbel', 'weibull', 'poisson', 'negativebinomial']
-crps_pgbm = np.zeros((len(tree_correlations), len(distributions)))
-for i, tree_correlation in enumerate(tree_correlations):
-    for j, distribution in enumerate(distributions):
-        print(f'Correlation {i+1} / distribution {j+1}')
-        model.params['tree_correlation'] = tree_correlation
-        model.params['distribution'] = distribution
-        yhat_dist_pgbm = model.predict_dist(X_val, n_forecasts=n_forecasts)
-        crps_pgbm[i, j] = model.crps_ensemble(yhat_dist_pgbm.cpu(), y_val).mean()
-        df_val = df_val.append({'method':method, 'dataset':dataset, 'fold':0, 'device':params['device'], 'validation_estimators': base_estimators, 'test_estimators':params['n_estimators'], 'rho': tree_correlation, 'distribution': distribution, 'crps_validation': crps_pgbm[i, j]}, ignore_index=True)   
+best_distribution, best_tree_correlation = model.optimize_distribution(X_val, y_val)
 #%% Train model on full dataset and evaluate base case + optimal choice of distribution and correlation hyperparameter
 # Set iterations to best iteration from validation set
 params['n_estimators'] = model.best_iteration + 1
@@ -94,15 +81,14 @@ params['n_estimators'] = model.best_iteration + 1
 model = PGBM()
 model.train(train_data, objective=objective, metric=rmseloss_metric, params=params)
 #% Probabilistic predictions base case
-model.params['tree_correlation'] = 0.03
 model.params['distribution'] = 'normal'
+model.params['tree_correlation'] = 0.03
 yhat_dist_pgbm = model.predict_dist(X_test, n_forecasts=n_forecasts)
 # Scoring
 crps_old = model.crps_ensemble(yhat_dist_pgbm.cpu(), y_test).mean()
 # Optimal case
-df_val_opt = df_val.loc[df_val.groupby(['dataset'])['crps_validation'].idxmin()]
-model.params['tree_correlation'] = df_val_opt[df_val_opt.dataset == dataset]['rho'].item()
-model.params['distribution'] = df_val_opt[df_val_opt.dataset == dataset]['distribution'].item()
+model.params['distribution'] = best_distribution
+model.params['tree_correlation'] = best_tree_correlation
 yhat_dist_pgbm = model.predict_dist(X_test, n_forecasts=n_forecasts)
 # Scoring
 crps_new = model.crps_ensemble(yhat_dist_pgbm.cpu(), y_test).mean()